requirements.txt
Django>=5.2,<6.0
mysqlclient>=2.2.0
python-dotenv>=1.0.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1 